Fix "full changelog" link
Prior to this commit the tags were hard coded in to the full changelog
link. This meant that it was basically useless as it was the same
(and incorrect) for each section.

This commit fixes that by ensuring that we storing the right information
in the datastructure that describes the changelog. This will need
more refactoring in the future, but for now it should be fine.