Improve viper config system
Prior to this commit, it was not clear how to configure this tool. There
were defaults set within the application but you needed to look
at the source code to find out what they were.

This commit improves the config system by writing out a config file
that contains the defaults supported by the application.