Migrate to GitHubs v4 API
Prior to this commit the new command made some assumptions about how
tags worked.

Annotated tags contain a tagger reference which has a pointer to when
the tag was created and also the commit that the tag points to.

Lightweight tags are just a pointer to a commit and therefore contain
no tagger information.

The assumption that lightweight tags also contained the tagger reference
meant that entries were not being added to the correct bucket.

This commit fixes that (hopefully) by using the v4 api to retrieve a
list of tag refs that are either lightweight or annotated. This allows
us to access the correct date depending on the type of ref that is
returned.