Implementation of next-version and unreleased
This commit adds support for an unreleased section and also adds
an optional next-version flag.

The unreleased section will display any qualifying entries that have
been added since the last tag.

The next-version flag allows the user to specify an arbitrary version
numbeer to use for the next release. If used, any qualifying entries
that have been added since the last tag will be grouped here and the
unreleased section will not be shown.