Add update check
This commit adds functionality to check for updates to the application.

The check runs as a persistent post execution task and will compare
the current version of the application with the latest version available
on github.

This check should never cause the application to fail so the method
simply returns a boolean value.