Add a command to view current config
Prior to this commit it was only possible to view config by inspecting
the config file created in the users home directory or the current
environment variables if any settings had been overridden.

This commit adds the `config` command that allows users to print the
current configuration of the application to stdout in either yaml
or json format.