Refactor githubclient package
This commit addresses a couple of different issues with this package.

Previously, everything was lumped in to `client.go`. This was fine
initially but it became harder to manage as the codebase grew.
Domains are now separated in by file.
Future work is probably needed to subdivide each domain in to it's own
package to prevent leakage.

Secondly, the methods for Tags and Pull were a bit scrappy and not very
efficient.
This was a left over from the "get it done" phase of the
project. These instances have now been resolved and I think the
implementation is closed to O(n) (maybe.. either way it's less garbage
than it was before).