Update entrypoint message
Prior to this commit the entry point for gh-changelog was fairly bland
and did not provide a great UX for first time users.

This commit is a step towards fixing that by ensuring that users have
enough information to achieve the bare minumum by just running the `gh
changelog` command without any subcommands.