Add new version flags
Prior to this commit it was not possible to scope the changelog to a
specific tag.

This commit adds two new flags that will allow a user to either specify
a tag to start from or build the changelog starting from the latest tag.

`--from-version`

This flag determines the version from which to start the changelog.
If the value passed does not exist as a tag, the changelog will
be built from the first tag.

`--from-last-version`

This flag will build the changelog starting from the last tag.
Using this flag will result in a changelog with one entry.