Move Builder in to a public package
This commit moves builder in to it's own public package. Additionally
the builder package has also been refactored. The previous builder
pattern has been dropped in favour of a standard options pattern.

Also methods have been broken down as much as possible with the hope
that it will increase readability and help reduce duplication in the
future.

Tests have also been updated, along with mocks, to reflect the new
structure.