Implement logging abstraction
This commit adds a new internal package called logging. It exposes the
Logger interface. The package also includes two loggers that implement
the Logger interface.

TEXT - this logger uses zerolog and will print out a sequential list of
events.

SPINNER - even though this is technically not a logger, this option
implements a spinner that is used to display events.

The type of logger used is determined by the LoggerType options. The
LoggerType is passed to the constructor which will return the correct
logging implementation.