Environment detection
This commit adds the environment package. It currently contains one
function that is used to determine whether or not the app is running in
CI.

For most popular CI products this assertion is made by including an
environment variable called `CI`.

IsCI checks for the existence of this environment variable. If it is
set to true then we return a boolean true.