Fix usage on repositories without tags
Prior to this change, if the tool was ran on a repo with no tags (a new
project for example) it would fail because there were no tags returned
by githubs api.

This change fixes that by updaitng the builder so that an error is
generated only if there are no tags and the --next-version flag has not
been provided.