Allow local configs
This commit adds the current working directory as a search path for
viper.

It also changes the name of the configuration file to `.changelog.yml`.

If there is a `.changelog.yml` file present in the current directory,
gh-changelog will prefer this over the one that is created inside
`.config/gh-changelog`.