Fix config initialization
Prior to this change, a .changelog.yml would always be created in the
current working directory.

The local config file should be optional and only used when users need
to override certain values.

To restore the original functionality we now use SafeWriteConfigAs to
ensure the config is placed in the expected directory on first run.

The ability to initialize local config should be fixed in #120.