Patch Version Parsing Code
In this change we are temporarily patching the code that handles version
parsing.

The code in package version has been borrowed from SemVer/v3/version and
slightly adapted to help this tool work arround an issue caused by
different pre-release delimeters.

The package can be removed when/if SemVer supports this.