Add missing line between change title and list
When a release has breaking changes, they are listed in a "Changed"
section.  Unlike other sections, this section title is directly followed
by the change list, while other sections have a blank line to separate
the title and the change list.

For consistency make sure we have that blank line after all titles.